requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
bitsandbytes==0.42.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
clip==1.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
ftfy==6.3.1
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.20.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
openai==1.86.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
